[Sema] Move AccessScopeChecker classes to separate file in AST.
